pyproject.toml
[project]
name = "dreamhouse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.8",
    "django-jazzmin>=3.0.1",
    "django-location-field>=2.7.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "redis==5.0.1",
    "six>=1.17.0",
]


[tool.isort]
profile = "hug"
skip = "build,.tox,.venv"
src_paths="isort,test"